refactor: index 컴포넌트 추가 및 네이밍 규칙에 맞는 컴포넌트명으로 수정
