+++ src/shared/ui/index.ts
@@ -17,3 +17,4 @@ export * from './sidebar'
 export * from './switch'
 export { default as Text } from './text/text'
 export * from './tooltip'
+export * from './toast'
