+++ src/shared/utils/index.ts
@@ -1,11 +1,15 @@
 export * from './cn'
 export * from './create-context-scope'
+
 export * from './Slot'
 export * from './Slottable'
 
 // Date
-export * from './isUndefined'
 export * from './formatDateToYMD'
 export * from './format-date-time-with-day'
 
 export * from './format-relative-date'
+
+export * from './dev-log'
+export * from './is-function'
+export * from './is-undefined'
